Implement server-side error-handling
